more commit  messages + usage. break out findProjectFromBlock to own file
